Migrate CATalog and multisig scheduling to absolute timestamps.
Consume the SDK timestamp state-scheduler API for CATalog CSV records,
launch, unroll (on-chain timestamp gating), and trusted full-schedule
verification. Update generic multisig sync/view messaging, reset test
schedules to 1/2/3, and apply minimal XCHandles compile casts until
ticket 05 owns that migration.

+++ src/cli/xchandles/verify_deployment.rs
@@ -372,7 +372,7 @@ pub async fn xchandles_verify_deployment(
             registration_period: record.registration_period,
         }
         .curry_tree_hash();
-        if record.block_height != block
+        if u64::from(record.block_height) != block
             || state.pricing_puzzle_hash != fph.into()
             || state.expired_handle_pricing_puzzle_hash
                 != XchandlesExpirePricingPuzzle::curry_tree_hash(
